requirements.txt
selenium
pytest
pytest-html
pyYAML
pylint
bandit